requirements.txt
streamlit
google-generativeai
Pillow
SpeechRecognition
PyAudio
Flask
requests
